Remove functions from static_vector that don't make sense
The (count, value) constructors are confusing at best, at worst they can
be ambiguous with the (iterator, iterator) overload. The same
functionality can be provided by a factory function, if required.

The swap implementation does the same thing as std::swap - since it
isn't any more efficient, we might as well not provide one and fall back
to the std one.

max_size() didn't make sense because the maximum size clearly is the
capacity.