Implement STRUCTURALIZE() macro
The macro takes a constant expression and converts it to a structural
type that is still fairly close to the original one.

For example, std::vector becomes std::array or std::tuple, depending
on whether all elements have the same type (imagine nested vectors - the
inner vectors need not have the same length, and therefore convert to
different instantiations of std::array - so the outer vector becomes a
tuple).